chore(mergeable): move MergeIdempotencyEntry from Domain/ to Internal/
The .EntityFrameworkCore package is a single-purpose data layer; a `Domain/`
folder housing a persistence-only bookkeeping entity (idempotency cache row)
muddled the layering. Relocates the type to the existing `Internal/` folder
alongside the DbContext, EfMergeService, and other persistence-only types.

Pure rename — no behavior change. All 34 Mergeable tests pass.
